+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Miscellaneous
 * New utility function `cut_quantile` for cutting a numeric vector into quantiles.
 * New utility function `cat_with_newline` for concatenating and printing with newline.
+* New check function `check_proportion` which checks for a single proportion.
 
 # hermes 0.1.1